Move auth error message to utils folder
